test: rename test support module
