+++ tests/prompt_resources/test_slot_schema_loader_json_view.py
@@ -16,7 +16,7 @@ if str(SRC_ROOT) not in sys.path:
 
 from a2a_t.common.prompt_resources.errors import PromptResourceParseError
 from a2a_t.prompt.common.models import PromptReference
-from tests.test_support import ManagedTempDirTestCase
+from tests.support import ManagedTempDirTestCase
 
 
 class SlotSchemaLoaderJsonViewTest(ManagedTempDirTestCase):
